+++ src/field/extension/arithmetic.rs
@@ -4,7 +4,7 @@ use super::*;
 /// ## Field operations
 
 /// Addition of two [`Ext`] elements.
-impl<const N: usize, const P: u32> Add for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Add for GaloisField<N, P> {
   type Output = Self;
 
   fn add(self, rhs: Self) -> Self::Output {
@@ -17,19 +17,19 @@ impl<const N: usize, const P: u32> Add for GaloisField<N, P> {
 }
 
 /// Addition assignment of two [`Ext`] elements.
-impl<const N: usize, const P: u32> AddAssign for GaloisField<N, P> {
+impl<const N: usize, const P: usize> AddAssign for GaloisField<N, P> {
   fn add_assign(&mut self, rhs: Self) { *self = *self + rhs; }
 }
 
 /// Sum of a collection of [`Ext`] elements.
-impl<const N: usize, const P: u32> Sum for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Sum for GaloisField<N, P> {
   fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
     iter.reduce(|x, y| x + y).unwrap_or(PrimeField::<P>::ZERO.into())
   }
 }
 
 /// Negation of an [`Ext`] element.
-impl<const N: usize, const P: u32> Neg for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Neg for GaloisField<N, P> {
   type Output = Self;
 
   fn neg(self) -> Self::Output {
@@ -41,14 +41,14 @@ impl<const N: usize, const P: u32> Neg for GaloisField<N, P> {
 }
 
 /// Subtraction of two [`Ext`] elements.
-impl<const N: usize, const P: u32> Sub for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Sub for GaloisField<N, P> {
   type Output = Self;
 
   fn sub(self, rhs: Self) -> Self { self + (-rhs) }
 }
 
 /// Subtraction assignment of two [`Ext`] elements.
-impl<const N: usize, const P: u32> SubAssign for GaloisField<N, P> {
+impl<const N: usize, const P: usize> SubAssign for GaloisField<N, P> {
   fn sub_assign(&mut self, rhs: Self) { *self = *self - rhs; }
 }
 
@@ -102,7 +102,7 @@ impl<const N: usize, const P: u32> SubAssign for GaloisField<N, P> {
 /// ## Inclusion of [`FiniteField`] operations
 
 /// Addition of a [`FiniteField`] element to an [`Ext`] element.
-impl<const N: usize, const P: u32> Add<PrimeField<P>> for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Add<PrimeField<P>> for GaloisField<N, P> {
   type Output = Self;
 
   fn add(mut self, rhs: PrimeField<P>) -> Self::Output {
@@ -112,12 +112,12 @@ impl<const N: usize, const P: u32> Add<PrimeField<P>> for GaloisField<N, P> {
 }
 
 /// Addition assignment of a [`FiniteField`] element to an [`Ext`] element.
-impl<const N: usize, const P: u32> AddAssign<PrimeField<P>> for GaloisField<N, P> {
+impl<const N: usize, const P: usize> AddAssign<PrimeField<P>> for GaloisField<N, P> {
   fn add_assign(&mut self, rhs: PrimeField<P>) { *self = *self + rhs; }
 }
 
 /// Subtraction of a [`FiniteField`] element from an [`Ext`] element.
-impl<const N: usize, const P: u32> Sub<PrimeField<P>> for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Sub<PrimeField<P>> for GaloisField<N, P> {
   type Output = Self;
 
   fn sub(mut self, rhs: PrimeField<P>) -> Self::Output {
@@ -127,12 +127,12 @@ impl<const N: usize, const P: u32> Sub<PrimeField<P>> for GaloisField<N, P> {
 }
 
 /// Subtraction assignment of a [`FiniteField`] element from an [`Ext`] element.
-impl<const N: usize, const P: u32> SubAssign<PrimeField<P>> for GaloisField<N, P> {
+impl<const N: usize, const P: usize> SubAssign<PrimeField<P>> for GaloisField<N, P> {
   fn sub_assign(&mut self, rhs: PrimeField<P>) { *self = *self - rhs; }
 }
 
 /// Multiplication of an [`Ext`] element by a [`FiniteField`] element.
-impl<const N: usize, const P: u32> Mul<PrimeField<P>> for GaloisField<N, P> {
+impl<const N: usize, const P: usize> Mul<PrimeField<P>> for GaloisField<N, P> {
   type Output = Self;
 
   fn mul(mut self, rhs: PrimeField<P>) -> Self::Output {
@@ -142,7 +142,7 @@ impl<const N: usize, const P: u32> Mul<PrimeField<P>> for GaloisField<N, P> {
 }
 
 /// Multiplication assignment of an [`Ext`] element by a [`FiniteField`] element.
-impl<const N: usize, const P: u32> MulAssign<PrimeField<P>> for GaloisField<N, P> {
+impl<const N: usize, const P: usize> MulAssign<PrimeField<P>> for GaloisField<N, P> {
   fn mul_assign(&mut self, rhs: PrimeField<P>) { *self = *self * rhs; }
 }
 ///////////////////////////////////////////////////////////////////////////////////////////////
